feat(tailwind): update navbar with tailwind
