+++ src/icons/Moon.svelte
@@ -1,0 +1,13 @@
+<script>
+    let clazz;
+    export { clazz as class };
+</script>
+
+<svg xmlns="http://www.w3.org/2000/svg" class={clazz} fill="none" viewBox="0 0 24 24" stroke="currentColor">
+    <path
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="2"
+        d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
+    />
+</svg>
