+++ src/icons/index.ts
@@ -1,0 +1,3 @@
+export * as Moon from './Moon.svelte';
+export * as Sun from './Sun.svelte';
+export * as User from './User.svelte';
