+++ src/routes/__layout.svelte
@@ -1,6 +1,6 @@
 <script>
     import '../app.css';
-    import Header from '$lib/Header.svelte';
+    import Header from '../components/Header.svelte';
     import Theme from '../components/Theme.svelte';
 </script>
